Panel fixed and is now focusable.
The old theme start show animation on obj show, if you want
disabled by default it doesn't work well. I changed the theme
default to be visible, don't need animation on "show".

And hidden_set was with inverted logic.

The focus highlight is in theme. And the widget react to
Return/Space.

SVN revision: 52814