Make it possible to use origin besides (0,0) for pan object.
Although some widgets are modified in this patch there should not be any visible
change. Actually this patch is intended to fix a Gengrid issue (patch following).

In Gengrid when the items were bigger than the grid itself the scroll would not
work correctly when the alignment was other than 0.0. That happened because the
smart scroller object was assuming the pan origin to be (0,0) and doing some
miscalculations based on that.



SVN revision: 54902