Hi Raster. This is Myungjae Lee.
 As I’ve talked to you just before, I’m sending you the patch file for
applying max size in sizing_eval function in the scroller.

This patch will work in the case below

-       elm_scroller_content_min_limit() function is set to 1 (either
width or height)

-       the content’s min size is growing up unlimitedly (such as
entry case)

 
 
 Then the min size of the scroller cannot exceed the max size of itself.
 
 (if we let it be grown up to the content’s min size, there is no way
to limit the size of the scroller and to enable scrolling.)

 
 
 Please consider this way to avoid failure case (min is greater than
max) in the sizing_eval function,

and give me a feedback if it doesn’t meet any other requirement.

 
 
 Thank you.
 
  
  
   
   
   P.S.) patch file was created based on Rev. 54766.



SVN revision: 54978