From: Jihoon Kim <jihoon48.kim@samsung.com> Subject: [E-devel] elm_map patch for supporting keyboard input
Most of elementary widgets can process keyboard input, but elm_map widget
has not. This patch makes elm_map be able to process keyboard input such as
Left, Right, Up, Down, PgDn, and PgUp key. In addition, map image will be
zoomed in or out when '+' or '-' key in keypad.

Please review it and let me know if there is any problem in this patch.
Merry Christmas!



SVN revision: 55840