From: Daniel Juyung Seo <juyung.seo@samsung.com> Subject: [E-devel]  [Patch] genlist item cache exception handling patch
I added "no cached item" handling.
In some case, items should not be cached and should not use caches.

For example, 
if there was a state change in item style, it should not be cached
because it remembers the state changes.
And if we use elm_genlist_item_item_class_update() API,
item should not be cached as well.
It needs to be a fresh item style without remembering the state changes.



SVN revision: 56585