From: Jaehwan Kim <jae.hwan.kim@samsung.com>
I found a problem in els_scroller. When the scroller is resized, it shows the location of 
the coordination which was saved into wanted_coordination. If you don't use the page,
it is no matter. But if you use the page function, the animation of the scroller is operated
by _smart_scrollto_x_animator, not by _smart_momentum_animator. 
In _smart_scrollto_x_animator, therer isn't _update_wanted_coordinates. So the value of
wanted_coordination doesn't update. I think _update_wanted_coordinates have to be added in
_smart_scrollto_x_animator and _smart_scrollto_y_animator.
Please check the attached patch file.



SVN revision: 64406