Elm test: Fixed label test's window size.
This is actually a pretty important thing to remember.
elm_win_resize_object_add resizes the content to the size of the window,
the size of the window is only applied when the window is shown, at least on
the ecore evas. Is that wanted? or should we take the evas-object geometry
there? The problem I'm worried about is: writing beyond buffer limits.

SVN revision: 64579