From: Minseok Kim <minseok3.kim@samsung.com> Subject: [E-devel] [E-deve] [Patch] Add new apis for editable set/get in multibuttonentry
Typically multibuttonentry is used as a composer.
But sometimes, Multibuttonentry can be used as a viewer.
Currently there is no way to use multibuttonentry as non-editable
viewer.
So I added new APIs which are elm_multibuttonentry_editable_set/get().
If editable is true, entry in multibuttonentry is activated and user can
compose item by pressing the return key.
If not, entry is hidden and muttibuttonentry works as a non-editable
viewer.



SVN revision: 73223