Elm: Fix ticket #1266: "delete,request" callback never gets called on window close
Delete requests where not being called when the frame close button was
pressed. I know this is not the 'ideal' fix (which is why I left
myself a FIXME here), HOWEVER the Ideal fix would mean breaking the
Feature Freeze (and potentially API), so I will wait until freeze is
over for that. For now, this fixes the 'bugger' in a non-instrusive way.



SVN revision: 74946