[elm] Rework focus_region and on_focus_region  hooks.
The former is now issued for all scrollables, with the behavior it had
before, and the latter is a new virtual base function, with a default
implementation leading to values matching the old case where one
didn't implement the hook.

Please help me test if everything is OK -- it seems so, here.



SVN revision: 75904