elementary/ctxpopup - [E-devel] [Patch] when parent of ctxpopup is resized, call dismissed signal
Before the patch, when window of device which has elementary rotates,
the ctxpopup doesn't emit the "dismissed" signal to application.
For this reason, if an application developer wants to make an event using "dismissed" signal when device rotates,
he/she cannot do anything.

But after add evas_object_smart_callback in the function "_on_parent_resize",
this problem is resolved.
That means, whether a device is rotated or not, when ctxpopup disappears, "dismissed" signal is emitted.
So I want you to check this patch.

Thank you.

Signed-Off-By: Bora Hwang<bora1.hwang@samsung.com>



SVN revision: 82918