elementary/widget - added widget orientation mode infra and 2 apis elm_object_orientation_mode_disabled_set(0, elm_object_orientation_mode_disabled_get()
Orientation Mode is used for widgets to change it's styles or to send signals
whenever it's window degree is changed. If the orientation mode is enabled
and the widget has different looks and styles for the window degree(0, 90,
180, 270), it will apply a style that is readied for the current degree,
otherwise, it will send signals to it's own edje to change it's any states if
the style doesn't be readied.



SVN revision: 83880