Adding an API to Dayselector.
This patches introduces a new API that will make the Dayselector more
flexible. Before this patch the only way that the week this was being shown
was according to the '%a' format type from the strftime() function,
However the user may want to abbreviate using another criterion. This patch
will allow the user to do that.