elc_fileselector: Add fileselector_double_tap_navigation_enable configuration
elm_filechooser used single tap not only for selection of file but also for
navigation of directory.
So it did not provide a way to select directory without navigation.

With fileselector_double_tap_navigation_enable option, single tap will be used
only for selection and double tap will be used for navigation of file or
directory.