Eo: fix macros for table
A col variable has been used inside the macro instead of the column
parameter. This first variable is from the legacy API and so didn't have
impact when the function was called from the legacy API.
In GUI Builder, the compilation of the generated C code was failing
because we used directly the Eo API and so was searching for the col
variable.