elm/win: Do not force maximized status.
When receiving a maximize event from the maximize button, do not force
sd->maximized. We must wait for a state change event, just like when a
window is maximized by the elm_win_maximized_set function. This will
ensure that a maximized smart event is sent.

Additionally, send a elm,state,maximized signal to the edje theme, so it
can update shadows and window decorations correctly.