elementary/widget - don't call theme_set() until base widget is not created yet.
On the widget creation time, we should not set the theme in the widget infra.
This will cause the corruption of the theme_hook working since the widget is not setup yet.
Also the theme_set() will be applied twice each widget will set the themes right after base widget is created.