elm entry: Fix anchor bug finding wrong anchor geometry.
Summary: Fix a bug with finding the proper geometry when reporting anchor
interaction. It's easy to see in entry anchor2 test in elementary_test
when clicking on the lower border of the anchor.

Anchor geometries are relative to the textblock object. The problem was
that this was accounted for using the edje object's geometry instead of
the textblock geometry thus causing an offset.

Reviewers: JackDanielZ, tasn

CC: seoz

Differential Revision: https://phab.enlightenment.org/D436

TAsn: Commit message edited by me.