entry: Add elm_entry_select_region_set API
Summary:
in case of GTK, it provides API to select a region of text (gtk_editable_select_region).
However, there are only API to select all or none in EFL.
This API will provide the way to select a region of text.

Test Plan: elementary_test > entry3 > press 'Region' button

Reviewers: woohyun, tasn, id213sin

Reviewed By: tasn

Differential Revision: https://phab.enlightenment.org/D520