Enhancement: #define names for text colors
Summary:
Previously text colors were specfied in the form color: XXX XXX XXX AAA throughout all the edc files, eg color: 255 255 255 255 this made it imposible to write a script to change text colors with sed as it changed many other parts of the theme as well.

Some colors had been defined in fonts.edc and in some places these were used but not many. This change means that all non textblock font colors are defined in fonts.edc and the defined names are used throughout the entire code base. This means that sed only needs to be run on 1 file to change the text color or it is easy to do manually

textblock colors were not done for 2 reasons
1. They hadn't been done in the past so there was no standard for me to follow.
2. They are all specified in html notation eg #FFFFFF which means that they can be changed with sed without effecting the rest of the theme.

If a standard can be come up with for doing this and people think it is worthwhile i can probably fix them at a later date

Reviewers: raster, zmike

Reviewed By: raster

CC: seoz

Differential Revision: https://phab.enlightenment.org/D577

Conflicts:
	data/themes/edc/battery.edc
	data/themes/edc/xkbswitch.edc