naviframe : Fixed the overlap issue during multiple push.
Summary:
Currently there is a overlap issue in naviframe in case of multiple push.
This issue will occur if the naviframe edc implements push,defered for
the item transition. The issue occurs becoz of cur,push,deffered is not being
called to the current and intermediate items during multiple item push.
To fix the above issue pass the data as current item that needs to be
out of the aviframe during creating the animator job.

@fix

Reviewers: Jaehyun, seoz, Hermet

Reviewed By: Hermet

CC: raster

Differential Revision: https://phab.enlightenment.org/D645