[Fileselector] Fix double-call of callback function on clicking "Enter" key.
Summary:
This feature (Clicking on "Enter" key) was previously implemented by using "activate" signal of entry (the one for entering name of the file),
so it was looking like "ok" button was clicked by pushing "Enter" key on the keyboard.
This little feature was added in commit 4d761f3a63270c88fa516764d9288f8bf0ff4a11.

Later the huge feature that adds key binding was implemented (commit 820f3dc907250c0c08ed8431b50567e58d6dab5c for ex).
It allows to bind widgets with keys (certain function calls when certain keys is pressed on keyboard).
This feature also adds same functionality to fileselecter, doing the same as previous (with "activate" signal) one.

Because of that it was looks like "ok" button was clicked twice.
This commit fixes this behavior and removes old and outdated method of "ok" button clicking processing.
@fix

Reviewers: cedric, seoz, raster

CC: reutskiy.v.v

Differential Revision: https://phab.enlightenment.org/D864