elementary: Don't segfault if we are not using an X11-compatible engine
When running Enlightenment under Wayland only, during init of E we
make a call to elm_config_all_flush. elm_config_all_flush is making
calls to ecore_x_window function(s) (for setting of elm_profile),
However this causes a crash if the ELM_ENGINE is not one that is
running under X11 (ie: ELM_ENGINE=drm or wayland_*). So to fix this
crash, we will compare the current ELM_ENGINE and see if it is
X11-compatible before making unnecessary (and crashing) calls to
ecore_x_window functions.

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>