Fix C++ box_cxx_example_02.cc and removed text_part_{set/get} from elm_button
As discussed in IRC, the commit that added text_part_set,
text_part_get functions to elm_button are wrong. The correct functions
to be used are from elm_layout: text_set and text_get.

Fixed the example to use these functions that are inherited on
elm_button from elm_layout and fixed the cleanup of childrens
in the parent window by using the del event callback.