configure: Use the same default CFLAGS/LDFLAGS as we do for EFL
Especially the -fvisibility=hidden would be needed here for our EAPI handling.
I never realised we miss this here but Lukasz Stanislawski pointed me to it
when we wondered why it linked ofr him with missing EAPI but not for me.

I'v compiling and releasing elm like this for a long time now so I don't
expect any trouble but if something strange comes up we can still revert it
before 1.11 final.