Revert "Revert "elm_notify: fixing style applying""
This reverts commit 2e2d003efeded27955cc9301279cd3a0d362f548.

Revert this patch to leave the credit to Andrii Kroitor.
I will fix this correctly in the next commit.
Sorry Andrii for the wrong review comments that can break compatibility.
But actually it was not documented at all so there was no problem.
Just to please hermet, let's not do more.