GLView: Add API to get the Evas_GL
Before screaming "don't expose this", here's the reasoning:
There will be a few new APIs in evas-gl (support pbuffer, ...)
that require a pointer to the Evas_GL to be called. So, instead
of exposing each and every one of these evas gl functions in
a dummy wrapper in elm_glview, we just give access to the real
pointer.
GLView will always be a wrapper around evas_gl, because that's
what it is by definition.