genlist: Fixed a crash bug in case of calling a genlist_clear in a select callback function.
Summary:
If item select callback is elm_genlist_clear and genlist object have many item, elm_genlist_clear was broken.
This reason is due to incorrect code. In Old code, check the item focus  after an item was deleted. So, I changed the order of the code.
Before item was deleted, Check the item focus.
@fix

Test Plan: I revised the elementary_test code and tested with that.

Reviewers: raster, SanghyeonLee, seoz

Reviewed By: seoz

Subscribers: bluezery, SanghyeonLee

Differential Revision: https://phab.enlightenment.org/D1418