Adapt del callback to Eo.
Because of widget item design, it is hard to make it fit to Eo as it
should be:
- _eo_del_cb should have the Evas_Smart_Cb as data but pre_notify
function prevents it because we need to be able to call the cb before
del is really triggered.
- only one cb is supported for an item, meaning replacing it would mean
we have to save the function pointer inside the item, delete the
callback with the previous function as data and then add the new
callback.