image: fix clipped image issue if x or y is less than zero
Summary:
After applying clipping patch about image on outside,
the width or height of the image be decreased
if x or y of an image is less than zero.
The way to calculate width/height has changed.

This fixes a side effect added in 2839881f37ea85b3469d8fd37cfaa4f9d67458fa

Reviewers: Hermet

Differential Revision: https://phab.enlightenment.org/D1810