genlist: send reorder signal to decoration item and views.
Previous logic was too much stick to special case for *tizen*
let decide item styles in theme as possible.

This might break reorder group item ui, but it won't be critical.
because tizen ui only depeneded on the scenario and they don't use them anymore.

This will be much better for multiple scenario in the future.