Elm_Entry: Remove explicit usage of Ecore selection
Selection handling is already handles at elm_cnp, so why does it still
exist in elm_entry? This is mostly problematic and confusing. My guess
is that the elm_entry code wasn't removed when it should've been when
elm_cnp was added.

TODO: continue this commit message with more info on what's going on.