Revert "win INLINE_IMAGE test now crashes as expected when trying to use it in a non-useless manner"
This reverts commit fd71874024d14e5374aef93eae7d4cfbc0ea4521.

As the commit message says, the test crashes now. The ticket 1973 is
related to that issue.
Now, I understand there could be a bug inside inline stuff but it doesn't
mean that you have to push it and screw up all the test. You could just
attach a patch to the ticket and that's all folks!