tests: actually run the main loop properly to be able to enter in idle state.
Note that the iterate function will never trigger the idler that are registered
in Ecore. That's by definition. I changed the code to actually use the full
main loop and trigger the change detection on idle enterer. That should be enough
for Elementary as all idler should logically affect the visual aspect of something
at some point and exit idle.

Thanks marcel-hollerbach@t-online.de for helping me debug this issue.