Icon: choose the fdo icon size using the evas obj geometry
Use the min between w and h of the evas object to
request the size in the icon theme. Fdo icon are always
square (1:1) so min seems correct.

Also improved the test to include a test that do
not relay on the size_hint_min of the icon.

Now in all the test I have done always the correct
fdo icon are used.