elm_image: Solve clash on the meaning of "animated"
Both Elm_Image and its base interface Efl.Image have a property named
"animated".
The meaning for Efl.Image is whether or not the image has animation.
On the other hand, Elm_Image use it to set whether an image, which supports
animation, is to animate itself or not.
Thus Elm_Image have another property (animated_available) to tell whether the
image have animation or not, making it even more confusing.

Renamed the "animated" property of Elm_Image to "animated_enable" to make
the intended meaning clearer.
Also replaced the "animated_available" property with the "animated" property
from the Efl.Image interface, since both refers to the same thing, in order
to reduce the API.

Added implementations of the legacy methods for backward compatibility.