elementary: Fix elm_win_window_id_get for wayland
Summary: This fixed the elm_win_window_id_get function when running in
wayland to return the ecore_wl_window_id rather than the surface id.
This change is necessary so that anyone calling this function can use
the return value in ecore_wayland function calls.

NB: Thanks to Hermet for pointing this out ! :)

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>