tests/build: Make sure tests are finding the theme file
Our tests require a theme file. So far they picked this just from the system
where it was installed to in a previous run. We have situations where this would
not work though. Running make check before install on a new system or during
distcheck comes to mind.

With AM_TESTS_ENVIRONMENT we make that the theme is also looked for in the not
yet installed builddir.

Thanks to Lukasz Stanislawski for pointing me to this.

ref T2028