First review
Hi Thiep,

I finally had time to review dnd. I put my comments in this patch.
One important thing is that we don't have to see the X11 implementation
as a strong reference. A lot of code is shitty there and we should throw
it on both implementations.
I didn't review all the code as I think there is a lot to fix already
and open issues.
Concerning the notify stuff, I don't think we should conserve the X11
way (i.e call cb inside notify or entry/image functions). We could,
for example, create data preparers that would be invoked before cb is
called.

Daniel