Theme: Fix SegmentControl border images & offsets
There was some bleeding on the images edges (with GL).
Changing the border values fixes that.

Also, fix TEXT offset difference when changing effects
(SHADOW --> GLOW). This means the text was moving when
selecting an item. The ellipsis had to be recalculated and
in many cases it would change, making one character appear
or disappear. Annoying.

Who uses this widget anyways?