NO PUSH: selection targets determination
This patch tries to improve the selection code by removing a function
that is totally wrong. This one checks data that is not some uri but
targets atoms ids sent by the source.
The problem is that now, during a cnp of text inside an entry, the requested
type is uri. The uri handler doesn't know how to deal with the sent
data. If we change it to retrive text when no file is found, the data
appears but is no more converted to plain text (markups characters).

A solution would be that when the uri handler fails, we request the next
type that was proposed by the source and supported by the target.