CnP: serialize requests to avoid inconsistency
When big data chunks are downloaded from an owner, parallel requests create
inconsistency and data become unusable.
This patch serializes the user requests by appending them into a waiting
list. When the list of types supported by the owner is received, the
request of the wanted type is handled in high priority.