elementary: Adjust shadow spacer on maximize
Summary: This fix adjusts the shadow spacer elementary when we
maximize/unmaximize a window. When in a maximized state, we do not
have to reserve space for the window shadow. This partially fixes
T2578. I say partially because there is still a little bit of blank
space around a maximized window.

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>