calendar: fix use of strftime.
strftime is quite a bad POSIX API. You won't know the size it needs to put the strings in
until you did give it enough byte and it could also fail without a clear way to detect it.
So this patch work around that and introduce a currently internal helper that will be moved
later to Eina.

This patch has been triggered by the issue reported and the attempt fix from D2877.