check: add "elm,activate,check,on" and "elm,activate,check,off" signal to support state change animation.
Summary:
Currently when user clicks on the checkbox, and when api state_check() called by
the developer we emit "elm,state,check,*" signal. To support state change animation when
user interacts with check box. There is no way to distinguish the action in EDC.
This "elm,activate,check,*" is a way which edc can make use to distinguish the stae change signal
by the user action or due to api call.

Reviewers: woohyun, raster, cedric, Hermet

Reviewed By: Hermet

Subscribers: id213sin

Differential Revision: https://phab.enlightenment.org/D2817

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>