build: Exclude some more files from coverage as they are not in the correct path
These files come from some modules in src/modules while we set out base dir to
src/lib. In result these files are search for in src/lib and not found. In the
long term we want to find out how to collect them as well but for now a working
coverage is already good.

Overall coverage rate:
  lines......: 18.1% (12596 of 69412 lines)
  functions..: 18.6% (1766 of 9500 functions)

Will work on getting these run by jenkins once it is back.