elm - thumbscroll finger scrolling - make it far smoother
smooth out scrolling and simplifyconfig as well as expose config apis
to configure it all and elm config ui sliders and checkboxes for
related config values etc. etc. - with this it is much smoother if you
use the default values, though latency is a bit worse. it looks nicer
though.

@feature