elementary: Fix calling ecore_evas_wayland_move with improper values
Summary: As ecore_evas_wayland_move ends up updating the location of
the ecore_wl_window, we should be calling move with the proper
geometry as this ends up updating the internal ecore_wl_window
geometry (which in turn sends to xdg_surface_set_window_geometry).
This is part of a fix where xdg_surface_set_window_geometry was
getting sent inproper values.

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>