elm_image: let's not trigger confusing update add.
Elm_Image rely on Evas_Image ability to rotate content. There is no
pixels manipulation involved that require an update_add here. In fact
this cause a bug where the content would disapear due to evas trying
to get some pixels data that are not where it think they are.