elm_test: Fix bug in GLView test
Scenario:
  const char * accel = accel_get();
  accel_set("gl:depth");

Now accel may not be valid anymore since the original
Eina_Stringshare might have been destroyed. So there was
no point in saving the const char * pointer. It was not valid
after the call to set(). This is a bit confusing here. Returning
char* would be clear: the caller would own the data.