notify: animate on hide
There was all the necessary code, except that the default
theme did not implement properly the signal & required data.

There is still a problem though, as the "timeout" smart callback
will be called at the start of the animation, so it's not possible
for an app to hide() or del() at this point (that would get
rid of the animation). But there's no other callback after
the animation ends.

elementary_test -to Notify > "Bottom" illustrates this issue.