win : Prohibiting auto-rendering, until elm_win is shown.
Summary:
Although elm_win is only created and evas_object_show with elm_win is not called,
evas_rendering works sometimes.
This rendering is not necessary. Because on one is shown and just back buffer is allocated.
This patch is the prohibiting auto-rendering, if elm_win is shown by calling evas_object_show()
It is effective for power consumption and reducing memory.

@fix

Test Plan:
1. elementary_test : checking all menus's working
2. modifed bg_example_02 : updated elm_bg's color by animator without evas_object_show(elm_win)

Reviewers: jpeg, jypark, raster

Reviewed By: raster

Differential Revision: https://phab.enlightenment.org/D3282