gengrid: fix Logic in _item_single_select_left and _item_single_select_right.
Summary:
In _item_single_select_left/right functions if logic is corrected from && to ||.
The previous thing is like: if "there is a generation mismatch" and "item is
disable" then only get the next gengrid item. This seems logically incorrect.

The suggested logic is if "there is generation mismatch" OR "item disabled" then
go for the new item. Whether we get a generation mismatch or item disable, we
should look for next item. So there should be || logic instead of && logic.

@fix

Signed-off-by: Umesh Tanwar <umesh.tanwar@samsung.com>

Reviewers: Hermet, raster, singh.amitesh, SanghyeonLee, cedric

Subscribers: cedric, sachin.dev

Differential Revision: https://phab.enlightenment.org/D3354

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>