Examples performance: Fix wrong always true condition.
This is again obviously wrong. This code tries to check if the type is not
one of three, what it does instead is checking if it's either not the
first two, or is the third, which is always true (because format can't
be too different values at the same time) and always wrong.

CID1327361