wayland cnp: check selection owner before calling losscb
The selection is cleared by losscb when we set it.
As result, selection is not done.
To avoid it, we should check if the new selection owner
is current selection owner or not and only call losscb
if the new selection owner is not the current one.

Test Plan:
In wayland, select text in entry (e.g double click, ctrl-a).