elm config - monitor JUST the correct files for profile and cfg change
this fixes the config monitoring to not use flush.cfg but to use the
correct profie.cfg or base.cfg in the profile we are using. this is
necessary to do "special" profiles per screen so apps that switched to
another profile because asked by e, monitor THAT profile for changes
specifically and dont also switch profiles since they have been told
what to use now.

@fix