combobox: move responsibility to define filter and content to the user of a combobox
Before this patch, the filter was set inside the combobox. This is confusing for the
user of the object as it can't define a custom filter except on strings. It is also
impossible to bind as binding require to know what was put in and they can't differentiate
between a genlist filter being set from the binding to one being set internally by
the combobox.

@fix