build: fix problems with coverage check build where some files are not found
Switch to use a lcov config file which geninfo_auto_base and remove hard coding
the base dir  to src/lib. geninfo_auto_base is designed for a use case like
ours where we have several base dirs (lib, bin, tests, ...) and it detects them
automatically. This fixes failures in a coverage run like this:

genhtml: ERROR: cannot read
/srv/build/workspace/manual_elm_gcc_x86_64_exactness_coverage/src/lib/test_win_state.c

Where the file is simply looked for in the wrong directory.