genlist: add elm_genlist_filtered_items_count API
Summary: This API will return the number of items that have passed the filter at a given point of time. If the API is called before "filter,done" callback comes, it may or may not give the filtered count for the complete genlist.

Test Plan: Updated Genlist_filter example to print the output from filtered_count_get API in "filter,done" callback.

Reviewers: SanghyeonLee, shilpasingh, cedric

Reviewed By: cedric

Subscribers: buds

Differential Revision: https://phab.enlightenment.org/D3725

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>