elm - genlist item - index get start from 0 - unbreak api break
so in feb 2015 seoz changed elm_genlist_item_index_get to start from 1
rather than 0. going back to elm code in 1.7 - it started at 0. this
is an api break that shouldn't have happend, but did. this fixes that.
yes - it's inconsistent with gengrid's index_get - but gengrid here is
wrong. nth_get starts at 0. this will get fixed with eo api's, and in
fact none of these index/nth api's should be in genlist's eo api.
legacy only. i can see why this was changed - it matches gengrid and
is more consistent, but we can't break things even if stupid.

@fix