interface_scrollable: fix the getting scroll view port size
Summary:
In auto scroller mode, both of vbar and hbar of scroller is created when scroller content size is getting large.
At the state, If we make hbar disappear by resizing scroller content, vbar size is set to small.
Because at the moment, the width and height of elm.swallow.content of sid->edje_obj are zero.

@fix

Test Plan:
1. set scroller auto mode
2. make vbar and hbar by resizing scroller content
3. make hbar disappear by resizing scroller content
4. see the vbar size whether it is correct or not

Reviewers: Hermet, Jaehyun_Cho, NikaWhite, cedric

Reviewed By: cedric

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D3779

Signed-off-by: Cedric Bail <cedric@osg.samsung.com>