Entry: Fix line_wrap logic
Calling single_line_set caused line_wrap to change to ELM_WRAP_NONE.
However, most of the logic prefers single_line over line_wrap e.g. when
updating the style of the widget. A few other places were missing this
check but that's pretty much it.

This fixes the following sequence of calls:
  elm_entry_line_wrap_set(ELM_WRAP_WORD);
  elm_entry_single_line_set(EINA_TRUE);
  elm_entry_single_line_set(EINA_FALSE);

@fix